Added a whole bunch of goodies.
Optparse is now working fully as far as I can tell. I would rather argparse but I'd like to limit us to Python 2.6 (EL6) as we may not be able to guarantee the presence of the builder.

Hooks are in place for periodic classad injection and exec wrappers.

Added a "realize_file" function that will download files over HTTP or copy files from a local path to the destination dir. This doesn't feel UNIXy enough yet but maybe I'll fix it later.